+++ cluster.lua
@@ -21,6 +21,7 @@ local vshard = require('vshard')
 local membership = require('membership')
 _G.vshard = vshard
 
+local vars = require('cluster.vars').new('cluster')
 local admin = require('cluster.admin')
 local webui = require('cluster.webui')
 local topology = require('cluster.topology')
@@ -28,15 +29,25 @@ local bootstrap = require('cluster.bootstrap')
 local confapplier = require('cluster.confapplier')
 local cluster_cookie = require('cluster.cluster-cookie')
 
-local e_init = errors.new_class('Cluster initialization failed')
+-- Parameters to be passed at bootstrap
+vars:new('box_opts')
+vars:new('boot_opts')
+vars:new('bootstrapped')
+
 --- Initialize the cluster module.
 -- After the call user can operate the instance via tarantool console.
 -- Notice that this call does not initialize the database - `box.cfg` is not called yet.
 -- The user must not try to call `box.cfg` himself, the cluster will do it when it's time to.
 -- @function init
--- @treturn true|nil Success indication
--- @treturn ?error Error description
+-- @treturn nil
+-- @raise
+--
+-- * `Can not create workdir`
+-- * `Missing port in advertise_uri`
+-- * `Socket bind error`
+-- * `Can not ping myself`
 local function init(opts, box_opts)
+    assert(vars.boot_opts == nil , 'Cluster is already initialized')
     checks({
         workdir = 'string',
         advertise_uri = 'string',
@@ -47,10 +58,10 @@ local function init(opts, box_opts)
 
     opts.workdir = fio.abspath(opts.workdir)
 
-    if not fio.path.exists(opts.workdir) then
+    if not fio.path.is_dir(opts.workdir) then
         local rc = os.execute(('mkdir -p \'%s\''):format(opts.workdir))
-        if not rc then
-            return nil, e_init:new('Can not create working directory %q', opts.workdir)
+        if rc ~= 0 then
+            error(('Can not create workdir %q'):format(opts.workdir))
         end
     end
 
@@ -73,22 +84,19 @@ local function init(opts, box_opts)
 
     local advertise = uri.parse(opts.advertise_uri)
     if advertise.service == nil then
-        return nil, e_init('Missing port in advertise_uri %q', opts.advertise_uri)
+        error(('Missing port in advertise_uri %q'):format(opts.advertise_uri))
     else
         advertise.service = tonumber(advertise.service)
     end
 
     log.info('Using advertise_uri "%s:%d"', advertise.host, advertise.service)
-    local ok, err = e_init:pcall(membership.init, advertise.host, advertise.service)
-    if not ok then
-        return nil, err
-    end
+    membership.init(advertise.host, advertise.service)
     membership.set_encryption_key(cluster_cookie.cookie())
     membership.set_payload('alias', opts.alias)
     -- topology.set_password(cluster_cookie.cookie())
     local ok, err = membership.probe_uri(membership.myself().uri)
     if not ok then
-        return nil, e_init:new('Can not ping myself: %s', err)
+        error(('Can not ping myself: %s'):format(err))
     end
 
     -- broadcast several popular ports
@@ -110,51 +118,80 @@ local function init(opts, box_opts)
     -- errors.monkeypatch_netbox_call()
     -- netbox_fiber_storage.monkeypatch_netbox_call()
 
-    local boot_opts = {}
-    boot_opts.workdir = opts.workdir
-    boot_opts.binary_port = advertise.service
+    vars.box_opts = box_opts
+    vars.boot_opts = {
+        workdir = opts.workdir,
+        binary_port = advertise.service,
+        bucket_count = opts.bucket_count,
+    }
 
     if #fio.glob(opts.workdir..'/*.snap') > 0 then
         log.info('Snapshot found in ' .. opts.workdir)
-        return bootstrap.from_snapshot(boot_opts, box_opts)
-    else
-        package.loaded['cluster'].bootstrap = function(roles, uuids)
-            checks('?table', {
-                instance_uuid = '?uuid_str',
-                replicaset_uuid = '?uuid_str',
-            })
-
-            local _boot_opts = table.copy(boot_opts)
-            _boot_opts.bucket_count = opts.bucket_count
-            _boot_opts.instance_uuid = uuids.instance_uuid
-            _boot_opts.replicaset_uuid = uuids.replicaset_uuid
-            return bootstrap.from_scratch(_boot_opts, box_opts, roles)
+        local ok, err = bootstrap.from_snapshot(vars.boot_opts, vars.box_opts)
+        if not ok then
+            log.error('%s', err)
         end
-
+    else
         fiber.create(function()
             while type(box.cfg) == 'function' do
-                if not bootstrap.from_membership(boot_opts, box_opts) then
+                if not bootstrap.from_membership(vars.boot_opts, vars.box_opts) then
                     fiber.sleep(1.0)
                 end
             end
-            package.loaded['cluster'].bootstrap = function()
-                return nil, e_init:new('Already bootstrapped')
-            end
+
+            vars.bootstrapped = true
         end)
         log.info('Ready for bootstrap')
     end
 
-    package.loaded['cluster'].init = function()
-        return nil, e_init:new('Already initialized')
-    end
-
     return true
+end
+
+local function bootstrap_from_scratch(roles, uuids)
+    assert(not vars.bootstrapped, 'Cluster is already bootstrapped')
+    checks('?table', {
+        instance_uuid = '?uuid_str',
+        replicaset_uuid = '?uuid_str',
+    })
+
+    local _boot_opts = table.copy(vars.boot_opts)
+    _boot_opts.instance_uuid = uuids.instance_uuid
+    _boot_opts.replicaset_uuid = uuids.replicaset_uuid
+
+    local function pack(...)
+        return select('#', ...), {...}
+    end
+    local n, ret = pack(
+        bootstrap.from_scratch(_boot_opts, vars.box_opts, roles)
+    )
+
+    vars.bootstrapped = true
+
+    return unpack(ret, 1, n)
+end
+
+--- Register user-defined role to be used in cluster.
+-- It should be done before calling `cluster.init()`
+--
+-- @function register_role
+-- @tparam string module_name A module to be loaded
+-- @treturn nil
+-- @raise
+--
+-- * All errors that `require` can raise
+-- * `Role "module_name" is already registered`
+-- * `Cluster is already initialized`
+local function register_role(...)
+    assert(vars.boot_opts == nil , 'Cluster is already initialized')
+    return confapplier.register_role(...)
 end
 
 return {
     init = init,
     admin = admin,
     webui = webui,
-    bootstrap = nil,
+    bootstrap = bootstrap_from_scratch,
     is_healthy = topology.cluster_is_healthy,
+
+    register_role = register_role,
 }